Update doctor Service class
