+++ doctor-service/src/main/java/com/healthcare_app/doctor_service/contoller/DoctorController.java
@@ -2,6 +2,7 @@ package com.healthcare_app.doctor_service.contoller;
 
 import com.healthcare_app.doctor_service.model.Doctor;
 import com.healthcare_app.doctor_service.service.DoctorService;
+import com.healthcare_app.doctor_service.service.DoctorServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
